main Window to view
